+++ src/domain/goals/GoalId.ts
@@ -1,0 +1,4 @@
+import { defineDomainIdentity, DomainIdentity } from "../identity/DomainIdentity.js";
+
+export type GoalId = DomainIdentity<"GoalId">;
+export const GoalId = defineDomainIdentity("GoalId");