+++ src/domain/workers/WorkerId.ts
@@ -1,0 +1,4 @@
+import { defineDomainIdentity, DomainIdentity } from "../identity/DomainIdentity.js";
+
+export type WorkerId = DomainIdentity<"WorkerId">;
+export const WorkerId = defineDomainIdentity("WorkerId");